backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.37
alembic==1.13.2
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
pydantic-settings==2.5.2
psycopg[binary]>=3.3,<3.4
python-multipart==0.0.9
pytest==8.3.3
httpx==0.27.2
twilio>=9.0.0
